refactor(core): move Mode to global Config
Mode was defined in models/input.py and carried as a field on the
Request model, forcing every backend that needed mode-awareness to
thread Request through its call stack.

Move Mode to a new hermeto/core/constants.py module and store it as a
field on the global Config object. This makes mode an execution-context
setting rather than domain data on Request, and lets any module query
the current mode via get_config().
The new module helps with keeping the import chain clean with no risk
of circular imports which we saw in the past.
