+++ hermeto/core/constants.py
@@ -1,0 +1,12 @@
+# SPDX-License-Identifier: GPL-3.0-only
+import enum
+
+
+class Mode(str, enum.Enum):
+    """Represents a global CLI option to relax input expectations and requirements checks."""
+
+    STRICT = "strict"
+    PERMISSIVE = "permissive"
+
+    def __str__(self) -> str:
+        return self.value
