+++ hermeto/interface/cli.py
@@ -14,10 +14,11 @@ import pydantic
 import typer
 
 from hermeto import APP_NAME
-from hermeto.core.config import set_config
+from hermeto.core.config import get_config, set_config
+from hermeto.core.constants import Mode
 from hermeto.core.errors import BaseError, InvalidInput, UnexpectedFormat
 from hermeto.core.extras.envfile import EnvFormat, generate_envfile
-from hermeto.core.models.input import Flag, Mode, PackageInput, Request, parse_user_input
+from hermeto.core.models.input import Flag, PackageInput, Request, parse_user_input
 from hermeto.core.models.output import BuildConfig
 from hermeto.core.models.sbom import Sbom, SPDXSbom, spdx_now
 from hermeto.core.resolver import (
@@ -191,7 +192,11 @@ def main(  # noqa: D103 -- docstring becomes part of --help message
 ) -> None:
     setup_logging(log_level)
     if config_file:
-        set_config(config_file)
+        config = set_config(config_file)
+    else:
+        config = get_config()
+    # Typer ensures `mode` is already a valid Mode enum value
+    config.mode = mode
 
 
 def _if_json_then_validate(value: str) -> str:
